+++ course/lib.php
@@ -614,40 +614,48 @@ function print_admin_links ($siteid, $width=180) {
     global $CFG;
     
     $icon = "<IMG SRC=\"$CFG->wwwroot/pix/i/settings.gif\" HEIGHT=16 WIDTH=16 ALT=\"\">";
-    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/config.php\">".get_string("configvariables")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/site.php\">".get_string("sitesettings")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/course/log.php?id=$siteid\">".get_string("sitelogs")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/theme/index.php\">".get_string("choosetheme")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/lang.php\">".get_string("checklanguage")."</A>";
-    $modicon[]=$icon;
-    if (file_exists("$CFG->dirroot/admin/$CFG->dbtype")) {
-        $moddata[]="<A HREF=\"$CFG->wwwroot/admin/$CFG->dbtype/frame.php\">".get_string("managedatabase")."</A>";
-        $modicon[]=$icon;
+    if (isadmin()) {
+	    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/config.php\">".get_string("configvariables")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/admin/site.php\">".get_string("sitesettings")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/course/log.php?id=$siteid\">".get_string("sitelogs")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/theme/index.php\">".get_string("choosetheme")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/admin/lang.php\">".get_string("checklanguage")."</A>";
+		$modicon[]=$icon;
+		if (file_exists("$CFG->dirroot/admin/$CFG->dbtype")) {
+            $moddata[]="<A HREF=\"$CFG->wwwroot/admin/$CFG->dbtype/frame.php\">".get_string("managedatabase")."</A>";
+			$modicon[]=$icon;
+		}
+		$moddata[]="<HR>";
+		$modicon[]="";
     }
-    $moddata[]="<HR>";
-    $modicon[]="";
-    $moddata[]="<A HREF=\"$CFG->wwwroot/course/edit.php\">".get_string("addnewcourse")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/course/categories.php\">".get_string("categories")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/course/teacher.php\">".get_string("assignteachers")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/course/delete.php\">".get_string("deletecourse")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<HR>";
-    $modicon[]="";
-    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/user.php?newuser=true\">".get_string("addnewuser")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/user.php\">".get_string("edituser")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/admin.php\">".get_string("assignadmins")."</A>";
-    $modicon[]=$icon;
-    $moddata[]="<A HREF=\"$CFG->wwwroot/admin/auth.php\">".get_string("authentication")."</A>";
-    $modicon[]=$icon;
+    if (iscreator()) {
+	    $moddata[]="<A HREF=\"$CFG->wwwroot/course/edit.php\">".get_string("addnewcourse")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/course/teacher.php\">".get_string("assignteachers")."</A>";
+		$modicon[]=$icon;
+    }
+    if (isadmin()) {
+	    $moddata[]="<A HREF=\"$CFG->wwwroot/course/categories.php\">".get_string("categories")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/course/delete.php\">".get_string("deletecourse")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<HR>";
+		$modicon[]="";
+		$moddata[]="<A HREF=\"$CFG->wwwroot/admin/user.php?newuser=true\">".get_string("addnewuser")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/admin/user.php\">".get_string("edituser")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/admin/admin.php\">".get_string("assignadmins")."</A>";
+		$modicon[]=$icon;
+        $moddata[]="<A HREF=\"$CFG->wwwroot/admin/admin.php\">".get_string("assigncreators")."</A>";
+		$modicon[]=$icon;
+		$moddata[]="<A HREF=\"$CFG->wwwroot/admin/auth.php\">".get_string("authentication")."</A>";
+		$modicon[]=$icon;
+    }
     $fulladmin = "<P><A HREF=\"$CFG->wwwroot/admin/\">".get_string("admin")."</A>...";
 
     print_side_block(get_string("administration"), "", $moddata, $modicon, $fulladmin, $width);
